Remove ExternalizableTypeElement
