+++ cr-examples/onnx/src/main/java/oracle/code/onnx/ir/OnnxType.java
@@ -25,14 +25,15 @@
 
 package oracle.code.onnx.ir;
 
-import jdk.incubator.code.extern.ExternalizableTypeElement;
+import jdk.incubator.code.TypeElement;
+import jdk.incubator.code.extern.ExternalizedTypeElement;
 import jdk.incubator.code.extern.TypeElementFactory;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-public abstract sealed class OnnxType implements ExternalizableTypeElement {
+public abstract sealed class OnnxType implements TypeElement {
 
     public static final TypeElementFactory FACTORY = new TypeElementFactory() {
         @Override
