+++ hat/examples/experiments/src/main/java/experiments/RawLayout.java
@@ -35,7 +35,6 @@ import java.lang.invoke.MethodHandles;
 import java.lang.reflect.Method;
 import jdk.incubator.code.*;
 import jdk.incubator.code.analysis.SSA;
-import jdk.incubator.code.extern.ExternalizableTypeElement;
 import jdk.incubator.code.extern.OpFactory;
 import jdk.incubator.code.dialect.core.CoreOp;
 import jdk.incubator.code.CodeReflection;
@@ -237,7 +236,7 @@ public class RawLayout {
     }
 
 
-    public static final class PtrType implements ExternalizableTypeElement {
+    public static final class PtrType implements TypeElement {
         static final String NAME = "ptr";
         final MemoryLayout layout;
         final JavaType rType;
