+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/TypeElement.java
@@ -1,6 +1,6 @@
 package jdk.incubator.code;
 
-import jdk.incubator.code.extern.ExternalizableTypeElement;
+import jdk.incubator.code.extern.ExternalizedTypeElement;
 
 /**
  * A type, that defines a set of values.
@@ -21,22 +21,9 @@ public non-sealed interface TypeElement extends CodeItem {
     /**
      * Externalizes this type element's content.
      *
-     * @implSpec
-     * The default implementation returns an externalized type element with
-     * an identifier that is the result of applying {@code toString} to this object, and
-     * with no arguments.
-     *
      * @return the type element's content.
      */
-    default ExternalizableTypeElement.ExternalizedTypeElement externalize() {
-        // @@@ Certain externalizable type elements are composed of other type elements,
-        // which may or may not be externalizable. OpWriter is designed to work with
-        // non-externalizable type elements, but in such cases OpParser will fail
-        // to parse what OpWriter produces.
-        // @@@ Should this throw UnsupportedOperationException
-        // @@@ Should this be a static helper method on ExternalizableTypeElement?
-        return ExternalizableTypeElement.ExternalizedTypeElement.of(toString());
-    }
+    ExternalizedTypeElement externalize();
 
     /**
      * Return a string representation of this Java type.
