+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/dialect/core/TupleType.java
@@ -1,6 +1,7 @@
 package jdk.incubator.code.dialect.core;
 
 import jdk.incubator.code.TypeElement;
+import jdk.incubator.code.extern.ExternalizedTypeElement;
 
 import java.util.List;
