+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/dialect/java/JavaRef.java
@@ -1,6 +1,6 @@
 package jdk.incubator.code.dialect.java;
 
-import jdk.incubator.code.extern.ExternalizableTypeElement;
+import jdk.incubator.code.TypeElement;
 
 /**
  * A symbolic reference to a Java class member or a Java type including members,
@@ -11,7 +11,7 @@ import jdk.incubator.code.extern.ExternalizableTypeElement;
  * reflected representation, much like the symbolic description of a Java type
  * can be resolved to an instance of {@link java.lang.reflect.Type Type}.
  */
-public sealed interface JavaRef extends ExternalizableTypeElement
+public sealed interface JavaRef extends TypeElement
         permits MethodRef, ConstructorRef, FieldRef, RecordTypeRef {
     // @@@ Enhance TypeElement to traverse children
     //     - Uniform tree traversal and transformation independent of
