+++ src/jdk.incubator.code/share/classes/jdk/incubator/code/dialect/java/PrimitiveType.java
@@ -26,6 +26,7 @@
 package jdk.incubator.code.dialect.java;
 
 import jdk.incubator.code.dialect.java.impl.JavaTypeUtils;
+import jdk.incubator.code.extern.ExternalizedTypeElement;
 
 import java.lang.constant.ClassDesc;
 import java.lang.invoke.MethodHandles.Lookup;
